fix: URI encoding of Matrix URIs and org handling
There is no well defined standard for if the : should be encoded or not.
As such we now check for both versions. Additionally orgPract is really
org+pract, so we need to check for both the individual values and their
combinations. All of this could be better documented in their respective
specifications.