fix: all contact API statements failing because "user" is a keyword on PG
The synapse db API seems to not quote table names. We could work around
that by quoting them ourselves when passing them as values, but that
might break in the future. Instead we just rename the column to
something less error prone.

Currently no migration is provided, since it is easier to rename the
table for now and let it be recreated. You may want to delete the old
table eventually.