chore: Add matrix team to CODEOWNERS
The matrix team is probably now the primary party responsible for
synapse and its modules. As such add us to the CODEOWNERS, so that we
can review stuff. Not removing workflows for now, since the matrix team
is quite small and it is very helpful, if sometimes external can do
reviews, when some of us are unavailable.