ci: properly write coverage files in CI
By default we were only running the coverage command from the default
environment, which doesn't write coverage files. The "cov" script in the
"ci" environment does.

Also extend that script to write coverage.xml and to the terminal in
addition to the lcov file, even though currently we only upload the
lcov.info.