feat: Purge rooms with no activity
Do this by finding the last actual message in a room and retrieving it's
time stamp. In the case that the room never had any messages(either
encrypted or clear text), use the room creation event. If it was longer
than the `grace_period` below, that room is scheduled for purging using
the infrastructure introduced in #43 .

Frequency of the scan itself is governed by `room_scan_run_interval`

New settings
```yaml
inactive_room_scan:
  enabled: true or false # an optional switch to disable processing of
this part of the room scan
  grace_period: 26w # An optional value for how long to keep a room
after the last eligible event,
                    # defaults to 6 months("26w"). Uses the 'duration
parsing' or a direct millisecond
                    # integer. Ignored if 'enabled' is false
```

New metric available for "Per-block Metrics" section:
* "get_timestamp_of_last_eligible_activity_in_room" will be found on the
Synapse worker responsible for background tasks