fix: Logging contexts slipping during startup federation list retrieval
Early in the startup process, the federation list is retrieved to check
that a server being booted up is an EPA server if it is configured to be
so. The coroutine from that process can throw an error, which then
messes up the logging context and shows a scary CRITICAL error. Fixed